Move everything related to the "online" feature to a dedicated module.
API change: the "verify" method has been moved from the Yubico struct to the crate root
